- renames RadioInput component to RadioInputGroup component - extract component from RadioInputGroup into RadioGroup component
