+++ src/signals/incident/components/form/RadioInput/index.js
@@ -1,68 +1,63 @@
-import React, { Fragment } from 'react';
+import React, { useCallback } from 'react';
 import styled from 'styled-components';
 import PropTypes from 'prop-types';
-import isObject from 'lodash.isobject';
-import { themeColor, RadioGroup, themeSpacing } from '@datapunt/asc-ui';
+import { useDispatch } from 'react-redux';
+import { Radio, Label } from '@datapunt/asc-ui';
 
-import Header from '../Header';
-import InputUsingDispatch from './InputUsingDispatch';
+import { resetExtraState, updateIncident } from 'signals/incident/containers/IncidentContainer/actions';
 
-const Info = styled.p`
-  color: ${themeColor('tint', 'level5')};
-`;
-
-const StyledHeader = styled(Header)`
-  margin-top: ${themeSpacing(5)};
-`;
-
-const RadioInput = ({ handler, touched, hasError, meta, parent, getError, validatorsOrOpts }) => {
-  const currentSelected = parent.meta.incident && parent.meta.incident[meta.name];
-  let info;
-  let label;
-
-  if (currentSelected && meta?.values[currentSelected.id]) {
-    ({ info, value: label } = meta.values[currentSelected.id]);
+const StyledLabel = styled(Label)`
+  & > * {
+    font-family: AvenirNextLTW01-Regular, arial, sans-serif;
   }
+`;
 
-  if (!meta?.isVisible) return null;
+const RadioInput = ({ checked, id, idAttr, label, info, name, resetsStateOnChange }) => {
+  const dispatch = useDispatch();
+
+  const onChange = useCallback(() => {
+    if (resetsStateOnChange) {
+      dispatch(resetExtraState());
+    }
+
+    dispatch(
+      updateIncident({
+        [name]: {
+          id,
+          label,
+          info,
+        },
+      })
+    );
+  }, [dispatch, id, info, label, name, resetsStateOnChange]);
 
   return (
-    <StyledHeader meta={meta} options={validatorsOrOpts} touched={touched} hasError={hasError} getError={getError}>
-      {meta.values && isObject(meta.values) && (
-        <Fragment>
-          <RadioGroup name={meta.name}>
-            {Object.entries(meta.values).map(([key, value]) => (
-              <InputUsingDispatch
-                checked={handler().value.id === key}
-                id={key}
-                idAttr={`${meta.name}-${key + 1}`}
-                info={value.info}
-                key={key}
-                label={value.value || value}
-                name={meta.name}
-                resetsStateOnChange={meta.resetsStateOnChange}
-              />
-            ))}
-          </RadioGroup>
-          {info && (
-            <Info>
-              {label}: {info}
-            </Info>
-          )}
-        </Fragment>
-      )}
-    </StyledHeader>
+    <StyledLabel htmlFor={idAttr} label={label}>
+      <Radio
+        checked={checked}
+        data-testid="inputUsingDispatch"
+        id={idAttr}
+        onChange={onChange}
+        type="radio"
+      />
+    </StyledLabel>
   );
 };
 
+RadioInput.defaultProps = {
+  checked: false,
+  info: '',
+  resetsStateOnChange: false,
+};
+
 RadioInput.propTypes = {
-  handler: PropTypes.func,
-  touched: PropTypes.bool,
-  getError: PropTypes.func.isRequired,
-  hasError: PropTypes.func.isRequired,
-  meta: PropTypes.object,
-  parent: PropTypes.object,
-  validatorsOrOpts: PropTypes.object,
+  checked: PropTypes.bool,
+  id: PropTypes.string.isRequired,
+  idAttr: PropTypes.string.isRequired,
+  info: PropTypes.string,
+  label: PropTypes.string.isRequired,
+  name: PropTypes.string.isRequired,
+  resetsStateOnChange: PropTypes.bool,
 };
 
 export default RadioInput;
