+++ src/signals/incident/components/form/index.test.js
@@ -6,7 +6,7 @@ import HandlingMessage from './HandlingMessage';
 import HiddenInput from './HiddenInput';
 import MultiTextInput from './MultiTextInput';
 import PlainText from './PlainText';
-import RadioInput from './RadioInput';
+import RadioInputGroup from './RadioInputGroup';
 import RedirectButton from './RedirectButton';
 import SelectInput from './SelectInput';
 import TextInput from './TextInput';
@@ -29,7 +29,7 @@ describe('Form components', () => {
       MapSelect,
       MultiTextInput,
       PlainText,
-      RadioInput,
+      RadioInputGroup,
       RedirectButton,
       SelectInput,
       TextInput,
