+++ src/signals/incident/definitions/wizard-step-2-vulaan/overlast-op-het-water.js
@@ -26,7 +26,7 @@ export default {
           nee: 'Nee',
         },
       },
-      render: FormComponents.RadioInput,
+      render: FormComponents.RadioInputGroup,
     },
     extra_boten_snelheid_rederij: {
       meta: {
